Closes #320
I also did a little bit of refactoring and made the Admin index page a bit more pleasant to look at.
