+++ admin/payoutadmin_sar_rescue.php
@@ -1,111 +1,28 @@
 <?php 
-// Mark all entry pages with this definition. Includes need check check if this is defined
-// and stop processing if called direct for security reasons.
+// REQUIRED on all secured pages
 define('ESRC', TRUE);
+require '../page_templates/secure_initialization.php';
 
-require_once '../class/db.class.php';
-include_once '../includes/auth-inc.php';
-require_once '../class/output.class.php';
-require_once '../class/users.class.php';
+// PAGE VARS
+$db = new Database();
+$user = new Users();
+$pgtitle = 'SAR Locate/Rescue Payouts';
 
-// if start and end dates are not set, set them to default values
-if (!isset($_REQUEST['start'])) {
-	$start = gmdate('Y-m-d', strtotime("- 7 day"));
-	$startPD = gmdate('Y-M-d', strtotime("- 7 day")); // formatted for Pikaday widget
-}
-if (!isset($_REQUEST['end'])) {
-	$end = gmdate('Y-m-d', strtotime("now"));
-	$endPD = gmdate('Y-M-d', strtotime("now")); // formatted for Pikaday widget
-}
 
-// set start and end dates to submitted values (GET or POST)
-if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
-	// start date
-	$arrStart = explode('-', $_REQUEST['start']);
-	$startYear = intval(substr($arrStart[0], -3)) + 1898;
-	$startMonth = intval(date('m', strtotime($arrStart[1])));
-	$startDay = intval($arrStart[2]);
-	$start = gmdate('Y-m-d', strtotime($startYear. '-' . $startMonth. '-' . $startDay));
-	
-	// end date
-	$arrEnd = explode('-', $_REQUEST['end']);
-	$endYear = intval(substr($arrEnd[0], -3)) + 1898;
-	$endMonth = intval(date('m', strtotime($arrEnd[1])));
-	$endDay = intval($arrEnd[2]);
-	$end = gmdate('Y-m-d', strtotime($endYear. '-' . $endMonth. '-' . $endDay));
-	
-	// special string for Pikaday widget
-	$startPD = htmlspecialchars_decode(date("Y-M-d", strtotime($startYear. '-' . $startMonth. '-' . $startDay)));
-	$endPD = htmlspecialchars_decode(date("Y-M-d", strtotime($endYear. '-' . $endMonth. '-' . $endDay)));
-}
-
+// HTML PAGE template - Begin
+require '../page_templates/home_html-begin.php';
 ?>
-<html>
 
-<head>
-	<?php
-	$pgtitle = 'Payout Admin';
-	include_once '../includes/head.php'; 
-	?>
-	<style>
-	<!--
-		table {
-			table-layout: fixed;
-			word-wrap: break-word;
-		}
-		a,
-		a:visited,
-		a:hover {
-			color: aqua;
-		}
-	-->
-	</style>
-	<script type="text/javascript">
-		$(document).ready(function() {
-		    $('#example').DataTable( {
-		        "order": [[ 0, "desc" ]],
-		        "pagingType": "full_numbers",
-		        "pageLength": 15
-		    } );
-		} );
-	</script>
-</head>
-
-<body>
-<div class="container">
-	<div class="row" id="header" style="padding-top: 10px;">
-		<?php include_once '../includes/top-left.php'; ?>
-		<div class="col-sm-8" style="text-align: center; height: 100px; vertical-align: middle;">
-			<span style="font-size: 125%; font-weight: bold; color: white;">Payouts: 
-				<a href="payoutadmin.php">ESRC</a> &gt;&gt; 
-				<a href="payoutadmin_sar.php">SAR Dispatch</a> &gt;&gt; 
-				SAR Locate/Rescue</span>
-			<span class="pull-right"><a class="btn btn-danger btn-md" href="index.php" role="button">
-				Admin Index</a></span><br />
-			<form method="post" class="form-inline" action="<?php echo htmlentities($_SERVER['PHP_SELF']); ?>">
-				<div class="input-daterange input-group" id="datepicker" style="margin-bottom: 5px;">
-					<input type="text" class="input-sm form-control" name="start" id="start" 
-						value="<?php echo isset($startPD) ? $startPD : '' ?>" />
-					<span class="input-group-addon">to</span>
-					<input type="text" class="input-sm form-control" name="end" id="end" 
-						value="<?php echo isset($endPD) ? $endPD : '' ?>" />
-				</div>
-				<div class="checkbox">
-					<label class="white"><input type="checkbox" name="details" value="yes"> Payout</label>
-				</div>
-				&nbsp;&nbsp;&nbsp;&nbsp;<button type="submit" class="btn btn-sm">Search</button>
-			</form>
-		</div>
-		<?php include_once '../includes/top-right.php'; ?>
+<div class="row">
+	<div class="col-sm-12" style="text-align: center; height: 100px; vertical-align: middle;">
+		<?php require '../page_templates/admin_payouts-header.php'; ?>
 	</div>
-	<div class="ws"></div>
-	<?php
-	// display results for the selected date range
-	$db = new Database();
-		
-	//show detailed records if "Payout" is not checked
-	if (!isset($_POST['details']) && $_POST['details'] != 'yes') {	
-	?>
+</div>
+
+<?php		
+//show detailed records if "Payout" is not checked
+if (!isset($_POST['payout'])) {	?>
+
 	<div class="row" id="systable">
 		<div class="col-sm-10">
 			<table id="example" class="table display" style="width: auto;">
@@ -135,16 +52,16 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 					echo '<td class="white text-nowrap">'. 
 							date("Y-m-d H:i:s", strtotime($value['LastUpdated'])+14400) .
 						 '</td>';
-					echo '<td><a target="_blank"
+					echo '<td><a class="payout" target="_blank"
 							href="https://evewho.com/pilot/'. $value['pilot'] .'">'.
 							Output::htmlEncodeString($value['pilot']) .'</td>';
-					echo '<td><a href="/esrc/rescueoverview.php?sys='. ucfirst($value['system']) .
+					echo '<td><a class="payout" href="/esrc/rescueoverview.php?sys='. ucfirst($value['system']) .
 							'" target="_blank">'. ucfirst($value['system']) .'</a></td>';
 					if (!empty($value['locateagent'])) {
 						echo '<td class="text-nowrap">
-							<a target="_blank" href="/esrc/personal_stats.php?pilot='.
+							<a class="payout" target="_blank" href="/esrc/personal_stats.php?pilot='.
 							urlencode($value['locateagent']) .'">'.
-							$value['locateagent'] .'</a> - <a target="_blank"
+							$value['locateagent'] .'</a> - <a class="payout" target="_blank"
 							href="https://evewho.com/pilot/'. $value['locateagent'] .'">EG</a></td>';
 					} 
 					else {
@@ -171,11 +88,13 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 			<?=gmdate('Y-m-d H:i:s', strtotime("now"))?> EVE<br /><br />
 		</div>
 	</div>
-	<?php
-	}
-	//show payout data if "Payout" is checked
-	else {			
-	?>
+
+<?php
+}
+
+//show payout data if "Payout" is checked
+else {			?>
+
 	<div class="row" id="systable">
 		<div class="col-sm-10">
 			<table class="table" style="width: auto;">
@@ -189,8 +108,6 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 				</thead>
 				<tbody>
 					<?php
-					// create instance object
-					$user = new Users();
 					//summary data
 					$db->query("SELECT rr.id, rr.locateagent, rr.system, 
 									datediff(rr.LastUpdated, rr.requestdate) AS daystosar, w.Class
@@ -215,23 +132,23 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 						// (base x WH class multiplier) + (Days until rescued x daily increase amt)
 						$payoutmax = ($basepay*$whclassmult)+(intval($value['daystosar'])*$dailyincrease);
 						echo '<tr>';
-						echo '<td><a target="_blank"
+						echo '<td><a class="payout" target="_blank"
 								href="/esrc/rescueoverview.php?sys='. ucfirst($value['system']) .'">'.
 								Output::htmlEncodeString(ucfirst($value['system'])) .'</a>
 								<span class="white">(50mil x '. 
 									intval(substr($value['Class'], -1)) .') + 
 									(10mil x '. intval($value['daystosar']).')
 								= '. number_format(intval($payoutmax)) .'</span></td>';
-						echo '<td><a target="_blank" 
+						echo '<td><a class="payout" target="_blank" 
 								href="https://evewho.com/pilot/'. $value['locateagent'] .'">'. 
 								Output::htmlEncodeString($value['locateagent']) .'</a></td>';
 						echo '<td>&nbsp;</td>';
 						// Locator gets half of total payout amount; if ESR Coord, they get 0
 						$payoutloc = intval($payoutmax/2);
-						$actualpayloc = ($user->isSARCoordinator($value['locateagent']) === false) ?
-							$payoutloc : 0;	
-						echo '<td><input type="text" id="amt'.$ctr.'" width="100" value="'. 
-								$actualpayloc .'" /><i id="copyclip" class="fa fa-clipboard" 
+						$actualpayloc = ($user->isSARCoordinator($value['locateagent'], false) === false) 
+							? $payoutloc : 0;	
+						echo '<td><input type="text" id="amt'.$ctr.'" width="100" 
+								value="'. $actualpayloc .'" /><i id="copyclip" class="fa fa-clipboard" 
 								onClick="SelectAllCopy(\'amt'.$ctr.'\')"></i></td>';
 						$totamt = $totamt + $actualpayloc;
 						echo '</tr>';
@@ -248,7 +165,7 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 								$ctr++;
 								echo '<tr>';
 								echo '<td></td><td></td>';
-								echo '<td><a target="_blank" 
+								echo '<td><a class="payout" target="_blank" 
 										href="https://evewho.com/pilot/'. $val['pilot'] .'">'. 
 										Output::htmlEncodeString($val['pilot']) .'</a></td>';
 								// first rescuer gets half of locator pay, then half again for each successive rescuer
@@ -280,80 +197,11 @@ if (isset($_REQUEST['start']) && isset($_REQUEST['end'])) {
 			<?php echo gmdate('Y-m-d H:i:s', strtotime("now"));?><br /><br />
 		</div>
 	</div>
+
 <?php
-	}
+}
+
+
+// HTML PAGE template - End
+require '../page_templates/home_html-end.php';
 ?>
-</div>
-
-<script type="text/javascript">
-	// datepicker
-    const monthNames = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
-
-    var startDate,
-    endDate,
-    updateStartDate = function() {
-        startPicker.setStartRange(startDate);
-        endPicker.setStartRange(startDate);
-        endPicker.setMinDate(startDate);
-    },
-    updateEndDate = function() {
-        startPicker.setEndRange(endDate);
-        startPicker.setMaxDate(endDate);
-        endPicker.setEndRange(endDate);
-    },
-    startPicker = new Pikaday({
-        field: document.getElementById('start'),
-        minDate: new Date('03/18/2017'),
-        showMonthAfterYear: true,
-        format: 'YYYY-MMM-DD',
-        toString(date, format) {
-            const day = ("0" + date.getDate()).slice(-2);
-            const month = monthNames[date.getMonth()];
-            const year = date.getFullYear() - 1898;
-            return `YC${year}-${month}-${day}`;
-        },
-        onSelect: function() {
-            startDate = this.getDate();
-            updateStartDate();
-        }
-    }),
-    endPicker = new Pikaday({
-        field: document.getElementById('end'),
-        minDate: new Date('03/18/2017'),
-        showMonthAfterYear: true,
-        format: 'YYYY-MMM-DD',
-        toString(date, format) {
-            const day = ("0" + date.getDate()).slice(-2);
-            const month = monthNames[date.getMonth()];
-            const year = date.getFullYear() - 1898;
-            return `YC${year}-${month}-${day}`;
-        },
-        onSelect: function() {
-            endDate = this.getDate();
-            updateEndDate();
-        }
-    }),
-    _startDate = startPicker.getDate(),
-    _endDate = endPicker.getDate();
-
-    if (_startDate) {
-        startDate = _startDate;
-        updateStartDate();
-    }
-
-    if (_endDate) {
-        endDate = _endDate;
-        updateEndDate();
-    }
-</script>
-
-<script type="text/javascript">
-	function SelectAllCopy(id) {
-	    document.getElementById(id).focus();
-	    document.getElementById(id).select();
-	    document.execCommand("Copy");
-	}
-</script>
-
-</body>
-</html>
